requirements.txt
autopep8==2.0.0
backoff==2.2.1
blinker==1.6.2
censys==2.2.0
certifi==2023.5.7
cfscrape==2.1.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cssselect==1.2.0
dnspython==1.15.0
Flask==2.3.2
html-similarity==0.3.2
idna==3.4
iniconfig==2.0.0
ipaddress==1.0.19
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parsel==1.2.0
Pillow==10.0.0
pluggy==1.2.0
pycodestyle==2.10.0
pydocstyle==6.3.0
Pygments==2.15.1
pytest==7.4.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
requests==2.31.0
rich==13.4.2
six==1.16.0
snowballstemmer==2.2.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.3
w3lib==2.1.1
websocket-client==1.5.2
Werkzeug==2.3.6
